rust-libc: allocator workload benchmarks
Adds bench/alloc.c (cargo xtask bench alloc): live-set churn at small,
medium and large sizes, realloc growth, tree build/teardown, a Larson
style multi-thread test, a producer/consumer ring and a fragmentation
RSS check, all compared against the host glibc.

+++ rust-libc/libc/src/signal.rs
@@ -78,7 +78,7 @@ fn valid(sig: c_int) -> bool {
 pub(crate) fn install_internal_handler(sig: c_int, handler: usize) {
     let act = KernelSigaction {
         handler,
-        flags: (crate::sys::SA_SIGINFO as u64) | SA_RESTORER,
+        flags: crate::sys::SA_SIGINFO | SA_RESTORER,
         restorer: restorer(),
         mask: 0,
     };
